ci: share common scripts utilities
